Rename project to Php Spreadsheet Wrapper
