+++ .gitignore
@@ -448,7 +448,8 @@ composer.phar
 
 
 # PHPUnit
-.phpunit.result.cache
+.phpunit.cache
+phpunit.xml
 
 # Commit your application's lock file https://getcomposer.org/doc/01-basic-usage.md#commit-your-composer-lock-file-to-version-control
 # You may choose to ignore a library lock file http://getcomposer.org/doc/02-libraries.md#lock-file
