+++ composer.json
@@ -1,5 +1,5 @@
 {
-  "name": "kign.org/php-spreadsheet-decorator",
+  "name": "kign.org/php-spreadsheet-wrapper",
   "description": "Convenience functions to PhpOffice/PhpSpreadsheet library",
   "type": "library",
   "license": "GPL-3.0-or-later",
