+++ src/PhpSpreadsheetWrapper/SpreadsheetLoader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace KignOrg\PhpSpreadsheetDecorator;
+namespace KignOrg\PhpSpreadsheetWrapper;
 
 use PhpOffice\PhpSpreadsheet\IOFactory;
 use PhpOffice\PhpSpreadsheet\Reader\Exception;
@@ -10,11 +10,11 @@ class SpreadsheetLoader
     /**
      * @throws Exception
      */
-    public static function loadFromFile(string $filename, bool $readDataOnly = false): SpreadsheetDecorator
+    public static function loadFromFile(string $filename, bool $readDataOnly = false): SpreadsheetWrapper
     {
         $inputFileType = IOFactory::identify($filename);
         $reader = IOFactory::createReader($inputFileType);
         $reader->setReadDataOnly($readDataOnly);
-        return new SpreadsheetDecoratorImpl($reader->load($filename));
+        return new SpreadsheetWrapperImpl($reader->load($filename));
     }
 }
