+++ src/PhpSpreadsheetWrapper/WorksheetWrapper/RowIteratorWithColumnNameImpl.php
@@ -1,6 +1,6 @@
 <?php /** @noinspection PhpPureAttributeCanBeAddedInspection */
 
-namespace KignOrg\PhpSpreadsheetDecorator\WorksheetDecorator;
+namespace KignOrg\PhpSpreadsheetWrapper\WorksheetWrapper;
 
 use Iterator;
 use PhpOffice\PhpSpreadsheet\Worksheet\RowIterator;
