+++ src/PhpSpreadsheetWrapper/WorksheetWrapper/WorksheetWithColumnHeaderImpl.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace KignOrg\PhpSpreadsheetDecorator\WorksheetDecorator;
+namespace KignOrg\PhpSpreadsheetWrapper\WorksheetWrapper;
 
 use InvalidArgumentException;
 use PhpOffice\PhpSpreadsheet\Cell\Cell;
